Fixing DrabTestApp loading
Altering the DrabTestApp to not be loaded if it does not exist, properly.

This whole case block and such really should not exist at all though, rather a proper mix test scaffold should be done that starts it on demand for testing, but this code should not even exist in hex releases.  This is the easiest work-around for now and is tested locally.