add function to get assigns from socket
Will fetch assigns from the priv section of socket, returning an empty list if none exist.

This is used when displaying the module examples when a user connects to a drab socket, so that their example is tailored to (one of) the assigns that they are relevant to the current socket.