Adds support for allowing PubSub subscription and events handling
PubSub subscription requires the PID of the main process, but in the
Commander the events callbacks are spawned asynchronously and a
temporary PID is assigned to them, different from the main one.
For this reason we need a synchronous callback to be called in the
Commander where to put the call to the PubSub subscription.
This new synchronous callback `onload_init`, is called just before
calling the  `onload` callback.

The GenServer `handle_info` PubSub events are handled calling
the `handle_info_message/2` callbacks, that is expected to be present
in the Commander for any different combination of topic/message
subscribed.

The `handle_info_message/2` callback in the Commander expect a
message parameter, identical to the standard `handle_info/2` one,
and a socket parameter, instead of the GenServer `state` parameter.
It haven't to return anything.

The original `state` parameter will be returned by the internal handling
of the GenServer/PubSub event.